refactor number to fhir
