+++ backend/src/main/java/gov/cdc/usds/simplereport/api/converter/FhirConverter.java
@@ -1,6 +1,12 @@
 package gov.cdc.usds.simplereport.api.converter;
 
+import com.google.i18n.phonenumbers.NumberParseException;
+import com.google.i18n.phonenumbers.PhoneNumberUtil;
+import com.google.i18n.phonenumbers.PhoneNumberUtil.PhoneNumberFormat;
 import org.apache.commons.lang3.StringUtils;
+import org.hl7.fhir.r4.model.ContactPoint;
+import org.hl7.fhir.r4.model.ContactPoint.ContactPointSystem;
+import org.hl7.fhir.r4.model.ContactPoint.ContactPointUse;
 import org.hl7.fhir.r4.model.HumanName;
 import org.hl7.fhir.r4.model.Identifier;
 import org.hl7.fhir.r4.model.Identifier.IdentifierUse;
@@ -29,5 +35,28 @@ public class FhirConverter {
       humanName.addSuffix(suffix);
     }
     return humanName;
+  }
+
+  public static ContactPoint phoneNumberToContactPoint(
+      ContactPointUse contactPointUse, String number) {
+    var contactPoint = new ContactPoint();
+
+    contactPoint.setSystem(ContactPointSystem.PHONE);
+
+    contactPoint.setUse(contactPointUse);
+
+    // converting string to phone format as recommended by the fhir format.
+    // https://www.hl7.org/fhir/datatypes.html#ContactPoint
+    try {
+      var phoneUtil = PhoneNumberUtil.getInstance();
+      var parsedNumber = phoneUtil.parse(number, "US");
+      var formattedWithDash = phoneUtil.format(parsedNumber, PhoneNumberFormat.NATIONAL);
+
+      contactPoint.setValue(formattedWithDash.replace("-", " "));
+    } catch (NumberParseException e) {
+      contactPoint.setValue(number);
+    }
+
+    return contactPoint;
   }
 }
