+++ backend/src/test/java/gov/cdc/usds/simplereport/api/converter/FhirConverterTest.java
@@ -2,13 +2,18 @@ package gov.cdc.usds.simplereport.api.converter;
 
 import static gov.cdc.usds.simplereport.api.converter.FhirConverter.convertToHumanName;
 import static gov.cdc.usds.simplereport.api.converter.FhirConverter.convertToIdentifier;
+import static gov.cdc.usds.simplereport.api.converter.FhirConverter.phoneNumberToContactPoint;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.from;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import org.hl7.fhir.r4.model.ContactPoint;
+import org.hl7.fhir.r4.model.ContactPoint.ContactPointSystem;
 import org.hl7.fhir.r4.model.Identifier.IdentifierUse;
 import org.hl7.fhir.r4.model.PrimitiveType;
 import org.hl7.fhir.r4.model.StringType;
+import org.hl7.fhir.r4.model.codesystems.ContactPointUse;
 import org.junit.jupiter.api.Test;
 
 public class FhirConverterTest {
@@ -54,5 +59,29 @@ public class FhirConverterTest {
       List<String> actual, List<StringType> expected) {
     var e = expected.stream().map(StringType::toString).collect(Collectors.toList());
     assertThat(actual).isEqualTo(e);
+  }
+
+  @Test
+  void mobileNumber_phoneNumberToContactPoint() {
+    var actual = phoneNumberToContactPoint(ContactPoint.ContactPointUse.MOBILE, "2485551234");
+
+    assertThat(actual)
+        .returns(ContactPointSystem.PHONE, from(ContactPoint::getSystem))
+        .returns("(248) 555 1234", from(ContactPoint::getValue))
+        .returns(
+            ContactPointUse.MOBILE.toCode(),
+            from(ContactPoint::getUse).andThen(ContactPoint.ContactPointUse::toCode));
+  }
+
+  @Test
+  void invalidNumber_phoneNumberToContactPoint() {
+    var actual = phoneNumberToContactPoint(ContactPoint.ContactPointUse.HOME, "+0333");
+
+    assertThat(actual)
+        .returns(ContactPointSystem.PHONE, from(ContactPoint::getSystem))
+        .returns("+0333", from(ContactPoint::getValue))
+        .returns(
+            ContactPointUse.HOME.toCode(),
+            from(ContactPoint::getUse).andThen(ContactPoint.ContactPointUse::toCode));
   }
 }
